+++ src/components/BulkEditList/BulkEditList.js
@@ -149,8 +149,10 @@ export const BulkEditList = () => {
           values={{ fileName: confirmedFileName || initialFileName }}
         />
       );
+    } else if (isQueryTabWithPreview) {
+      return <FormattedMessage id="ui-bulk-edit.meta.query.title" />;
     } else return <FormattedMessage id="ui-bulk-edit.meta.title" />;
-  }, [confirmedFileName, initialFileName, isIdentifierTabWithPreview, history.location.search]);
+  }, [confirmedFileName, initialFileName, isIdentifierTabWithPreview, isQueryTabWithPreview]);
 
   const changedPaneSubTitle = useMemo(() => {
     if (!isIdentifierTabWithPreview && !isQueryTabWithPreview) return null;